Fix a Fuchsia formatter type mismatch flagged by the pending Clang roll (flutter/engine#43651)

+++ engine/src/flutter/shell/platform/fuchsia/runtime/dart/utils/vmo.cc
@@ -69,7 +69,8 @@ bool VmoFromFilename(const std::string& filename,
       fdio_open_fd(filename.c_str(), static_cast<uint32_t>(flags), &fd);
   if (status != ZX_OK) {
     FX_LOGF(ERROR, LOG_TAG, "fdio_open_fd(\"%s\", %08x) failed: %s",
-            filename.c_str(), flags, zx_status_get_string(status));
+            filename.c_str(), static_cast<uint32_t>(flags),
+            zx_status_get_string(status));
     return false;
   }
   bool result = VmoFromFd(fd, executable, buffer);
@@ -91,7 +92,8 @@ bool VmoFromFilenameAt(int dirfd,
                                              static_cast<uint32_t>(flags), &fd);
   if (status != ZX_OK) {
     FX_LOGF(ERROR, LOG_TAG, "fdio_open_fd_at(%d, \"%s\", %08x) failed: %s",
-            dirfd, filename.c_str(), flags, zx_status_get_string(status));
+            dirfd, filename.c_str(), static_cast<uint32_t>(flags),
+            zx_status_get_string(status));
     return false;
   }
   bool result = VmoFromFd(fd, executable, buffer);
